tracing: tracing: Add functions to inject code to library files
As mentioned in the previous commit, to trace the flow of execution of
kw we need to inject code into the main kw executable (that was done in
the previous commit) and into library files (almost all files in `src`).

With that in mind, add three functions `sync_kw_lib_files_with_tracing`,
`inject_lib_file_with_tracing` and `get_tracing_log_line` to
`tracing/tracing.sh` that, combined, take care of injecting code that
enables tracing into all necessary files and synchronizes these with kw
local library (usually `~/.local/lib/kw`).

The code injected is the one that actually logs every entry and exit of
a function defined in the kw project. In other words, for every file
injected with code, we have:

- The first command of every function is to log the entrance in it;
- Before every `return` and `exit` statement in a function, there is a
  command to log the exiting of it;
- Before the end of a function definition (closing curly braces), there
  is a command to log the exiting of it preserving the return value of
  the last command;
- For every command that launches a background execution - which spuns
  another thread of execution - there is setup and teardown code (before
  and after the command) to separate the log of different threads.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>