src: ui: patch_hub: patch_hub_core: add new warning message_box
When you select the option "Registered Mailing list" without selecting
any mailing list, the UI prints a blue background without any
information for the user.

In this sense, update the code to show an "Error" message for the user
with instructions to Register a Mailing list before checking that
subject.

Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: JGBSouza <joaosouzaaa12@usp.br>