src: env: Block env switch if the kernel tree is not clean
One of the requirements of using kw env is imposed by how the O= option
works, which requires that the kernel tree is clean to work properly.
Currently, the command kw env --use only allows users to switch if the
env is clean, which can lead to build failures and consequently mislead
users. This commit addresses this issue by adding a check phase before
the user tries to switch to a new env; if the kernel tree is not clean,
kw fails the operation with an informative message.

Closes: #705

Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>