integration_tests: Vagrantfile: Introducing Integration Testing to the kw build
This commit implements support for integration tests for the `kw build` command.

We use Vagrant to provision a virtual machine and Ansible to configure it.
The purpose of this integration test is to create a virtual machine and execute
the `kw build` command within it to verify that all steps occur as expected.

Signed-off-by: Aquila Macedo <aquilamacedo@riseup.net>