src: device_info: Add verbose mode
This commit enables the verbose option for the device info;
unfortunately, to make these changes happen, I had to find a balance
between updating legacy code and adding a new one. In particular, I had
to create a new verbose function only for this feature because most of
the code processes strings, and the verbose option should display them
in the logs. Additionally, I decided to drop many tests that validate
the command sequence, which, for the specific case of device_info, is
not helpful because the potential of the errors are in the command
result and test it is the correct approach.

Related-to:  #857

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>