kw: show notice message if kw runs in developer mode
When kw runs in developer mode, some env vars are changed to point the
cache/data/db/doc/etc/lib/man directories to the repository in which the
`kw` script is at.

This can cause confusion to developers, which may expect some files
(such as database files, manual pages,  plugin files) to be somewhere
else and may think kw has a bug (due to devmode) when it doesn't.

This commit basically shows a notice message to inform the developer
that kw is running in developer mode and some envvars have been changed.
The message is sent to stderror to prevent failing tests (because it
would disrupt the output expected to the commands).

Signed-off-by: uwla <uwlalinux@gmail.com>