src: lib: dialog: Move `handle_exit` to `patch_hub_core.sh`
The function `handle_exit`, implemented in `src/lib/dialog_ui.sh`, is
intended to be used only inside the infinite while loop of the function
`patch_hub_main_loop`, implemented in
`src/ui/patch_hub/patch_hub_core.sh`, which it is.

For this reason, move the implementation of `handle_exit` to the latter
file, as it makes more logical sense for it to be there.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>