src: ui: patch_hub: patch_hub_core: Revise documentation of `handle_exit`
The function `handle_exit` is used by the patch-hub feature to
gracefully stop the FSM that coordinates the execution of the feature.
The feature is design using the MVC design pattern, and `handle_exit` is
only used by the Controller component (files of `src/ui/patch_hub`).

The documentation of the function misleads the reader to think that it
used by Dialog (the View component), so revise documentation.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>