src: ui: patch_hub: patch_hub_core: Make `show_dashboard` use `handle_exit`
The function `handle_exit` is used to stop the execution of the
patch-hub feature. In the Dashboard, if the user selects the 'Exit'
button, `show_dashboard` prints a error message that is not displayed to
the user and return, instead of calling `handle_exit`. This is both
outdated and violates the pattern accross the files in `src/ui/patch_hub`.

In this sense, make `show_dashboard` use `handle_exit` and delete unit
test that does not makes sense anymore.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>