src: deploy: Stop using mkinitcpio preset file
The deploy code has some references for the mkinitcpio.preset file
spread around the code. Recently, Melissa pointed out that we don't need
that file and we can use a single command:

mkinitcpio --generate /boot/initramfs-<kernel-name>.img --kernel <kernel-name>

This command represents a considerable simplification in the deploy
code, and for this reason, this commit makes it part of this change.

Closes #882

Tested-by: Melissa Wen <melissa.srw@gmail.com>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>