kw: show notice message on repo mode
When kw runs in repo mode (script was called  from  KW  repository,  not
from local installation), some kw env vars  are  changed  to  point  the
cache/data/db/doc/etc/lib/man directories to the repository in which the
`kw` script is at.

This can cause confusion to developers,  which  may  expect  some  files
(such as database files, manual pages, plugin  files)  to  be  somewhere
else and may think kw has a bug (due to devmode) when it doesn't.

This commit basically shows a notice message  to  inform  the  developer
that kw is running in repo mode and some envvars have been changed.

Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: uwla <uwlalinux@gmail.com>