tests: move unit tests to tests/unit, add flag --unit to target them
Reestructuring the tests folder by moving unit tests to tests/unit  will
make KW more organized and prepared for the incoming integration tests.

This commit also adds the flag `--unit` to `run_tests`, which limits the
scope of the tests to unit test. This may not seem like a big  deal  yet
becayse there are only unit tests, but it  will  be  relevant  once  the
integration tests are incorporated.

Signed-off-by: uwla <uwlalinux@gmail.com>