src: config: distinguish variable sources (global/local) in output
The `config --show` command now prints a header specifying the source of
variables along with them.

This requires the test to adapt to the new output: ignore  the  keywords
GLOBAL/LOCAL in the output and expect the default scope to be empty.

A note about the new default value  for  scope:  If  scope  defaults  to
local, the configuration output would  only  show  local  variables.  If
defaults to global, it would only show global variables. But if defaults
to empty, we can know the user has not specified an scope and  therefore
we should show both global and local configs.  That's  why  its  default
value is now an empty string.

Closes: #894

Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: uwla <uwlalinux@gmail.com>