src: lib: kw_config_loader: fix not reading line if missing newline char
The `kw_config_loader` module would not read the option in the last line
of a file if the line is missing the newline character '\n'.

While it is a good practice to always add the newline character in  text
files, KW should still work in cases where it is missing, perhaps due to
a mistake. Indeed, `kw init` creates config files some of which  do  not
have the newline character in the last line (such as build.config) which
is evidence that mistakes may happen, but even in such scenarios we want
KW to function properly.

Signed-off-by: uwla <uwlalinux@gmail.com>