src: lib: lore: Add pagination support to mailing lists
The 'Registered/Unregistered Mailing Lists' screen has some missing
mailing lists, due to the pagination of lore.kernel.org response.

For example, the mailing list 'lvm-devel' doesn't appear in patch-hub
because it is in the second page of lore.kernel.org.

So, add pagination factor to display all the available mailing lists.

Issue: #901

Signed-off-by: André Gustavo Nakagomi Lopez <andregnl@usp.br>
Co-authored-by: Joao Paulo Pereira da Silva <jppaulo11@hotmail.com>