setup.sh: add option to install kw system wide
KW can now be installed system wide, allowing multiple users to run kw.

For the system-wide installation of KW to work properly, the help module
was slightly changed to correct a  bug  which  prevented  manpages  from
being read in a system-wide installation (the bug did not occur in  home
installation).

Closes #985

Signed-off-by: uwla <uwlalinux@gmail.com>
Co-authored-by: João Gabriel Josephik <joao.gabrielaaj@gmail.com>