src: help.sh: show correct version on repo mode
When running on repo-mode, KW  did  not  show  up  the  correct  version
because it was simply running `cat` on a  file.  This  works  in  normal
installations because the `kw` installed version is copied to  a  static
file during the installation. In repo-mode such information is  dynamic,
not static. In this case, it is necessary to run other commands.

Also, call `kworkflow_function` in `setup.sh` to avoid code duplication.
Before, `setup.sh` would run the same commands as in `src/help.sh`.

Furthermore, add a unit test for kw version on repo mode.

Closes #986

Signed-off-by: uwla <uwlalinux@gmail.com>
Co-authored-by: João Gabriel Josephik <joao.gabrielaaj@gmail.com>