run_tests: allow running tests with the same basename
The run_tests script used a `strip_path` function which would store  the
basename of all files in the global variable  TESTS.  When  running  the
tests, we would use the find command to get the path of a file  matching
the basename of the current tests, then would check if that file  exists
and run the tests.

The problem with this approach is that it assumes all tests have  unique
basename. This is no longer the case: we now have  the  `config_test.sh`
integration test in addition to the `config_test.sh` unit test, and  the
same for device test. The old approach caused errors when attempting  to
run integration and unit tests with the same file

Running `./run_tests.sh --integration test device` would work.
Running `./run_tests.sh --unit test device` would work.
Running `./run_tests.sh test device` would fail.

To prevent this, this commit replaces the `strip_path` function with the
`set_tests` function, which stores the path of the test instead  of  its
basename. When running the tests, we use the `basename` command  on  the
test path to print out the basename of the  current  test,  rather  than
getting the path of the test from the basename of the test.

Also, the new approach uses regex to match specified tests when  running
`./run_test.sh test <test1> <test2>`. That is because the  find  command
cannot handle complex regex, only wildcards, so we use grep's  to  match
multiple test files via the provided regex.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@risup.net>
Signed-off-by: uwla <uwlalinux@gmail.com>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@risup.net>