tests: unit: help: add tests for kw --version
Add unit tests for kw version under normal installation.

Also, the repo-mode message is now sent to stderr. That way, it  can  be
easily ignored with a simple  redirection  to  /dev/null.  This  can  be
useful in future integration  tests  which  may  run  kw  in  repo-mode.
Otherwise, the output would always be messed up by the repo-mode message
and have to be filtered out in every test.

Signed-off-by: uwla <uwlalinux@gmail.com>