documentation: dependencies: fedora: Fix procps dependency in fedora
In fedora distro, when installing kw with setup.sh, it is always
requested to install the package procps, even when it's already
installed.

The problem is that the command `rpm -q procps` seems to specifically
not recognize that this package is already installed. However, with `rpm
-q procps-ng` (which is the same exact package) it works fine.

So, change the dependency to 'procps-ng' instead of 'procps' to fix this
problem.

Fixes: #1010

Signed-off-by: João Paulo Pereira da Silva <jppaulo11@hotmail.com>