tests: integration: Fix permission denied inside containers due to SELinux
SELinux is causing permission troubles to podman containers when the
current mode is set to "enforcing". It does not allow to execute
setup.sh inside the container, for example.

This problem is related with how podman is built, being, currently, a
quite common podman issue. Yet, it has an easy fix: add :z when
mounting the volume.

Helps: #1016

Signed-off-by: João Paulo Pereira da Silva <jppaulo11@hotmail.com>