src: ui: patch_hub: patchset_details_and_actions: Fix returning from 'Details and Actions' bug
[What]

With the implementation of 'Search String in Lore', the function
`show_latest_patchsets_from_mailing_list` accepts an argument for
additional filters to be passed in the Lore request. This argument is
passed via the `screen_sequence` data structure used to pass information
between the screens that compose patch-hub UI. However, because in the
screen 'Details and Actions' the
`screen_sequence['SHOW_SCREEN_PARAMETER']` contains data about a patchset
(not an additional filter) and because this value isn't reset when
returning from this screen, `show_latest_patchsets_from_mailing_list`
makes an invalid request that causes the fetch to fail, when it
shouldn't.

In this context, make the screen 'Details and Actions' reset the value
of `screen_sequence['SHOW_SCREEN_PARAMETER']` when returning and adapt
other parts to not also break the 'Search String in Lore' feature.

[Why]

Undoubtedly, users will go back and forth between the listing of latest
patchsets and a patchset details and actions, so this bug is a major
flaw in patch-hub.

[How]

The fix goes by simply resetting the value of
`screen_sequence['SHOW_SCREEN_PARAMETER']` in the function
`show_patchset_details_and_actions` when the user hits the button
'Return'. However, because 'Search String in Lore' also uses this value
to work, just this reset breaks this feature. So, besides the reset,
we add a global variable to patch-hub UI declared in `patch_hub_core.sh`
to pass additional filters instead of using
`screen_sequence['SHOW_SCREEN_PARAMETER']` for this.

Closes: #1028

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>