src: lib: lore: Make query filter for Lore request more robust
When fetching patches from Lore, patch-hub composes a query URL that
contain base filters along with additional filters. The base filters
basically aim to filter out any message from the given mailing list that
is not a patch. However, this query filter simply filters out messages
that have the string 'Re:' in the subject.

In this context, make this query filter more robust by enforcing that
the subject mustn't contain the string 're:', while also containing the
strings 'PATCH' or 'RFC' (all case insensitive). The query filter is
also constructed as an isolated logical expression to not conflict with
potential added filters.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>