src: lib: lore: Revise display of patchset information
When viewing both latest patchsets from a mailing list and the
bookmarked patchsets, the display of each item in the listing can be
improved. First, the message title has no size limit and overflows the
screen breaking the formatting of Dialog for long titles. Also, as the
the received time in Lore servers of the patch and the author name are
available, we can display these infos for the latest patchsets.

In this context, make the display of the patchset message title limited
to 60 chars, and display the received time and author name infos in the
listing of latest patchsets from a given mailing list.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>