src: lib: dialog_ui: Remove `LC_CTYPE` override for async loading screen spinner
The function `create_async_loading_screen_notification` creates a Dialog
loading screen with a spinner to signal users that a delayed action is
happening underneath. However, users reported environments that had
problems with the characters used for the spinner while using `kw
patch-hub` and running the test suite for `src/lib/dialog_ui.sh`.

The problem stems from the environment variable `LC_CTYPE` (used to
determine character handling rules) being overridden to account for
using the spinner characters that are Unicode (multibyte). The logic
behind this is sound, as the encoding aspect should be considered
because these spinner characters have a length of 3 bytes. Nevertheless,
not handling this matter fixes the problem for the environments
mentioned above and doesn't introduce regressions for the environments
that didn't face the issues.

In this context, remove the override of `LC_CTYPE` and consider the
spinner characters of length 1.

Closes: #965

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>