src: remote: Improve ssh handling in case of failure
As part of the kernel workflow, it is common for a developer to have
multiple hard disks with different distros for checking the kernel
behavior in those scenarios. When the developer swaps the disk but the
IP is still the same for the target machine, it is expected to hit the
remote host identification issue. This commit alleviates this problem by
detecting and taking action based on this type of failure.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>