src: build: Remove redundant override of `load_build_config`
In kw, the `src/lib/kw_config_loader.sh` module takes care of
configuration precedence for all the configurations, i.e., it composes
configurations from the most preferred scope to the least one. Although
the function `load_build_config` defined inside
`src/lib/kw_config_loader.sh` takes care of this behavior for the
configurations of `kw build`, this function is also defined in
`src/build.sh`, doing virtually the same as the former definition. This
double definition causes the one from build to override the one in the
config loader (config loader is sourced first).

Three problems arise from this situation:
 - Duplicated/redundant code.
 - The definition from build has a bug that outputs a warning to the
   user that doesn't makes sense.
 - It breaks the project's pattern of delegating any and all
   configuration loading to the `src/lib/kw_config_loader.sh` module.

Probably, `load_build_config` from build was created before the
inception of the config loader and we forgot to remove it after adding
the module. That may be the reason why the code is duplicated and has a
bug that doesn't exist in the config loader version.

To address the three problems described above, remove the definition of
`load_build_config` from `src/build.sh`.

Fixes: #824

Thanks-to: Rodrigo Siqueira <Rodrigo.Siqueira@amd.com>
Thanks-to: André Gustavo Nakagomi Lopez <andregnl@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>