src: kw_remote: Avoid destroying symbolic links in the add operation
If the user tries to add the same remote name inside an env, kw will
destroy the symbolic link and replace it with an actual file. Next, if
the user switches between envs, they lose the previous configuration.
This commit addresses this issue by adding the --follow-symlinks in the
add operation, which avoids symbolic link replacement.

Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>