src: _kw: Revise `kw remote` Zsh completions
The Zsh completions for the `kw remote` feature are broken, because:

  1. The `--list`, `--add`, `--rename`, and `remove` aren't mutually
     exclusive, as they should.
  2. The `--add` and `--rename` options require 2 arguments, but the
     completions assume they don't need any and suggest other options.
  3. The `--remove` option requires 1 argument, but the completions
     assume it doesn't need any and suggest other options.
  4. The `--set-default|-s` option followed by `--add` shouldn't accept
     arguments, yet it does.

In this sense, fix problems 1 to 3 by declaring `--add`, `--rename`, and
`--remove` as completions to `_arguments` instead of `_describe`, and
fix 4 by altering the completion of `--set-default|-s`, if `--add` has
been completed.

Also, to further improve the readability of the code, add names to the
arguments of options that need them (see the option `--add`, for
reference). These are not obligatory but are a way of documenting the
arguments.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>