tests: integration: utils: fix container_exec for enhanced command execution
This adjustment, including proper handling of options, container names, and the
command to execute with special consideration for the '/bin/sh -c' option, promotes
a more reliable and secure execution process. The commit aims to enhance the overall
functionality and robustness of the container_exec function when dealing with various
command scenarios.

Closes: #1040

Signed-off-by: Aquila Macedo <aquilamacedo@riseup.net>