src: lib: lore: Handle patch not yet processed when electing representative
When electing representative patches in the `kw patch-hub` feature,
which is vital to correctly listing patchsets, there is a border case
that is not consider: when the possible representative patch has
a in-reply-to message that has not yet been processed. In this case, the
logic always determine that the patch is representative, which isn't
true every time.

To mitigate this, make a request for the raw version of the in-reply-to
message that hasn't been processed, extract the necessary information,
and determine if the original patch is representive as before.

This implementation adds a network requisition that can incur in the
same problem `kw patch-hub` had of many chained requisitions. However,
as the number of requests is predictable (one per patch, in the extreme
case), when we add parallelization to the processing, this will be
solved.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>