src: lib: lore: Parallelize `process_representative_patches`
The processing of patches for listing patchsets has two major stages
after fetching the Atom feed with the patches and pre processing it. The
second major stage consists in traversing the processed list of patches
sequentially and determining if they are or not representative. This
stage is represented by the `process_representative_patches`.

As we can leverage parallel computing in this case, parallelize the
function by issuing a dedicated thread for handling each patch. This
takes care of much of the processing done in
`process_representative_patches` resulting in a gain of performance,
specially because it solves the problem of possible many chained network
requests.

Helps: #911

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>