src: lib: lore: Obtain patch metadata more robustly
The `kw patch-hub` feature depends heavily on correctly attaining
a patch metadata. This includes the patch version, the total number of
patches in its series, and its title stripped from the patch tag
(strings in the fashion of `[PATCH vX YY/ZZ]`, and the like).

However, the version and the patch tag are obtained by parsing the patch
title in not the most reliable way. The case is worst when getting the
total number of patches in the series, as this triggers chained network
requests that can range around one hundred in extreme cases, resulting
in unsatisfactory (and unpredictable) performance.

To make the attainment of patch metadata more robust, do the following:

  1. Extract methods to get each type of metadata (this was all mixed
     inside the function `extract_metadata_from_patch_title`);
  2. Consider more cases when parsing the title for metadata;
  3. Remove any network logic when obtaining the total number of
     patches.

Note: This commit also adds a function to get the number/index of the
patch in the series metadata, which is not being used by `kw patch-hub`
at the moment but will be in future commits.

Helps: #911

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Co-authored-by: André Gustavo <andregnl@usp.br>
Co-authored-by: Guilherme Moreno <guilhermewr9@gmail.com>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>