src: lib: lore: Overhaul processing of representative patches
[What]

Revise the logic in processing representative patches for listing
patchsets. When listing patchsets (a set of related patches designed to
form a more significant change), a patch is elected to represent it
(either a cover letter or the actual first patch in the series), and is
called the representative.

[Why]

Correctly defining a patch as a representative or not is vital for the
`kw patch-hub` feature. Having false negatives (representative patches
that are "invisible" to the feature) shouldn't happen and also impact
the performance, as deeper fetches for data are needed for the same
number of patchsets. Currently, in some lists like git, `kw patch-hub`
sometimes misses 80% of the time, which is unacceptable.

[How]

The heuristics to determine a patch as representative is flawed because
it solely depends on the pattern of message IDs. A consistent pattern
doesn't seem to exist across the many lists that can be reliably used to
determine this information (actually, message ID patterns aren't even
consistent in the same list).

At the moment, from fetching an Atom feed with a list of patches to
having a list of representative patches sorted from the newest to the
oldest is done by:

  1. Pre-processing the Atom feed to obtain only the needed attributes
     of each patch;
  2. Traversing the pre-processed list of patches once and deciding if
     a patch is representative based on its message ID.

We chang step 1 to include more attributes and replace step 2 with two
steps to overhaul this processing and make it more robust and reliable.

After pre-processing the Atom feed, we traverse the list of patches,
marking everyone in a hashtable and storing their processed versions on
an array. Then, we traverse the list of patches again, but their
processed versions, and use a combination of heuristics based on the
patch metadata and its relation with its in-reply-to message (if it has
one). These heuristics are inspired by the source code of the b4
project, which can be found at https://github.com/mricon/b4.

Note: Many files had to be updated due to ripple effects in the change
of variable names, and the like. Also, much of the size of this commit
is due to testing code and mocked samples for them.

Closes: #900

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Co-authored-by: André Gustavo <andregnl@usp.br>
Co-authored-by: Guilherme Moreno <guilhermewr9@gmail.com>