src: lib: lore: Parallelize `process_individual_patches`
The processing of patches for listing patchsets has two major stages
after fetching the Atom feed with the patches and pre processing it. The
first major stage consists in traversing the pre processed list of
patches sequentially, processing each one of them individually, and
marking everyone in a hashtable. This stage is represented by the
`process_individual_patches`.

As we can leverage parallel computing in this case, parallelize the
function by issuing a dedicated thread for handling each patch. This
takes care of much of the processing done in
`process_individual_patches` resulting in a gain of performance.

Helps: #911

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>
Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>