src: lib: kw_db: Update database handling functions for WHERE clauses
The original databases function had some limitations, as that
the select function couldn't handle a where clause to select an
specific data and the where clause used in the remove function
could only handle equality comparassions.
In this sense, add a function that generate where clause with
multiple relational operations and update the `select_from` and
`remove_from` functions to use this new implementation.

Reviewed-by: David Tadokoro <davidbtadokoro@gmail.com>
Signed-off-by: JGBSouza <joaosouzaaa12@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@gmail.com>