src: lib: kw_db: Update database handling functions
The original databases function had some limitations, as that
the select function couldn't handle a where clause to select an
specific data and the where clause used in the remove function
could only handle equality comparassions, also, there was no
easy method for update a data attribute without replacing the
whole line.
In this sense, add a function that generate where clause with
multiple relational operations and update the `select_from` and
`remove_from` functions to use this new implementation and
implements other function `update_into` that updates an specific
attribute of the database entry. Also change all
the select_from calls and add test cases for all the new cases.

Signed-off-by: JGBSouza <joaosouzaaa12@usp.br>