src: report: fix the default error messages
If the kw report is called without any arguments
or with --all flag and the statistics and pomodoro
data is not present then ideally we should get two
separate errors. But the option_value['ERROR']
value is getting overwritten. So we get two lines
displaying the no data for pomodoro error.
This commit adds an if clause for this condition
and modifies the tests accordingly.

Signed-off-by: Sahil Sagwekar <sahilss2652@gmail.com>