tests: unit: lib: lore_test: Replace TEST-MODE with TEST_MODE
The unit test test_get_raw_lore_message hangs because it needs an
internet connection, which is not expected for a unit test. This error
happened because of a typo in the unit test flag, which originally used
TEST-MODE instead of TEST_MODE. This commit addresses this issue by
using the correct flag.

Fixes: 9c4a5dd ("src: lib: lore: Handle patch not yet processed when electing representative")
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>