kw: replace every "#!/bin/bash" shebang with "#!/usr/bin/env bash"
The usual "#!/bin/bash" may not work for some linux distributions.
Replace it with "#!/usr/bin/env bash"

kworkflow, being a bash-written project, requires the use of shebangs to
define the interpreter environment where kw is run. Nowadays, the
scripts contain the shebang "#!/bin/bash", which works normally for most
of the distros.  There are, however, some exceptions, such as NixOS,
where there is no /bin/bash (bash is located in /nix/store). Using
workflow on these types of distros may not be smooth, since the user is
required to run kw commands preceded by bash or manually modify the
code.

Closes: #1089

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: Marcelo Mendes Spessoto Junior <marcelomspessoto@gmail.com>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>