src: send_patch: Rename `kw mail` to `kw send_patch`
This commit changes the name of the 'kw mail' feature to 'send-patch'.
The motivation behind this change is that the name 'kw mail' was too
generic and did not accurately represent the functionality of the
feature.

The 'kw mail' feature is designed to send patches via email, and its
name should reflect this. The new name, 'send-patch', is more
descriptive and immediately gives users an understanding of what the
feature does.

This change is important for improving the user experience and
making the functionality of the features clear at first glance.

Signed-off-by: Lucas Quaresma <lucasqml08@usp.br>
Co-authored-by: Roberto Bolgheroni <robertobolgheroni@usp.br>