src: plugins: kernel_install: utils: clean coding style violations
Replace short options with long options, which are more readable.
Replace echo with printf, which is more consistent across multiple
platforms than echo.

Note: The maintainance of cat instead of using Bash redirects (which is
the correct way considering the project coding style) is because
redirects don't support glob expansions, so using it instead of cat
would actually make the code more complex, and, possibly, less
performatic in the end.

Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: Fernando Yang <hagisf@usp.br>
Signed-off-by: Eduardo Figueredo <eduardofp@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>