tests: integration: utils: Add function for executing commands in nested containers
This commit introduces a new function container_exec_in_nested_container()
to execute commands within a container that is itself running inside
another container.

It takes the name or ID of the outer container, the name or ID of the
inner container, the command to be executed within the inner container,
and any additional parameters for podman container exec. The function
properly constructs the command string and executes it using
container_exec(). It also handles errors and returns the output of the
command if successful.

Signed-off-by: Aquila Macedo <aquilamacedo@riseup.net>