src: create build_and_deploy file
Create build_and_deploy utility to centralize kw bd functionalities.

Even though currently kw bd is only a "wrapper" for executing
kw build and kw deploy, in the future it can have its own options.
Therefore, this commit adds a file for kw bd along with a man page for
it, making it easier for future enhancements to the command.

Reviewed-by: Rodrigo Siqueira <siqueirajordao@riseup.net>
Signed-off-by: Briza Mel Dias <brizamel.dias@gmail.com>
Co-authored-by: Lorenzo Bertin Salvador <lorenzobs@usp.br>
Signed-off-by: Rodrigo Siqueira <siqueirajordao@riseup.net>