src: lib: kwlib: Change is_a_patch function verification
The current check to determine if a file is a patch done by kw
checked for text patterns in the files, checking if their
formatting was similar, which, however, could lead to certain
errors in certain cases. However, as git has a command
`git apply check`, which serves to check whether or not a
patch is applicable to a given repository, also serving to
check whether the file is patched or not, it replaces the
current verification method with the command, making scanning
more consistent.

Signed-off-by: JGBSouza <joaosouzaaa12@gmail.com>